Initial Implementation
I have added the main, product_cache, Create_tables files along with a few placeholder HTML files. This is to provide a basic understanding of the program's direction. I have also included a function that creates a database if one does not exist already using the Create_tables.sql file.